requirements.txt
Flask==2.1.2
pandas==1.4.2
PyMuPDF==1.19.6
python-dotenv==0.19.2
ibm-watsonx-ai  # Ensure the correct version is specified if needed
